+++ packages/pkg-error/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @business-as-code/error
 
+## 0.0.0-bollards-20230520145558
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520145106
+
+### Patch Changes
+
+- new message
+- new message
+
+## 0.0.0-bollards-20230520144145
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520143640
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520143126
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520142811
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520142731
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520142605
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230520142317
+
+### Patch Changes
+
+- new message
+
 ## 0.0.0-bollards-20230520081808
 
 ### Patch Changes