+++ packages/plugin-dev-changesets/CHANGELOG.md
@@ -1,5 +1,106 @@
 # @business-as-code/plugin-dev-changesets
 
+## 0.0.0-bollards-20230520145558
+
+### Patch Changes
+
+- new message
+- Updated dependencies
+  - @business-as-code/address@0.0.0-bollards-20230520145558
+  - @business-as-code/core@0.0.0-bollards-20230520145558
+  - @business-as-code/error@0.0.0-bollards-20230520145558
+  - @business-as-code/fslib@0.0.0-bollards-20230520145558
+
+## 0.0.0-bollards-20230520145106
+
+### Patch Changes
+
+- new message
+- new message
+- Updated dependencies
+- Updated dependencies
+  - @business-as-code/address@0.0.0-bollards-20230520145106
+  - @business-as-code/core@0.0.0-bollards-20230520145106
+  - @business-as-code/error@0.0.0-bollards-20230520145106
+  - @business-as-code/fslib@0.0.0-bollards-20230520145106
+
+## 0.0.0-bollards-20230520144145
+
+### Patch Changes
+
+- new message
+- Updated dependencies
+  - @business-as-code/address@0.0.0-bollards-20230520144145
+  - @business-as-code/core@0.0.0-bollards-20230520144145
+  - @business-as-code/error@0.0.0-bollards-20230520144145
+  - @business-as-code/fslib@0.0.0-bollards-20230520144145
+
+## 0.0.0-bollards-20230520143640
+
+### Patch Changes
+
+- new message
+- Updated dependencies
+  - @business-as-code/address@0.0.0-bollards-20230520143640
+  - @business-as-code/core@0.0.0-bollards-20230520143640
+  - @business-as-code/error@0.0.0-bollards-20230520143640
+  - @business-as-code/fslib@0.0.0-bollards-20230520143640
+
+## 0.0.0-bollards-20230520143126
+
+### Patch Changes
+
+- new message
+- Updated dependencies
+  - @business-as-code/address@0.0.0-bollards-20230520143126
+  - @business-as-code/core@0.0.0-bollards-20230520143126
+  - @business-as-code/error@0.0.0-bollards-20230520143126
+  - @business-as-code/fslib@0.0.0-bollards-20230520143126
+
+## 0.0.0-bollards-20230520142811
+
+### Patch Changes
+
+- new message
+- Updated dependencies
+  - @business-as-code/address@0.0.0-bollards-20230520142811
+  - @business-as-code/core@0.0.0-bollards-20230520142811
+  - @business-as-code/error@0.0.0-bollards-20230520142811
+  - @business-as-code/fslib@0.0.0-bollards-20230520142811
+
+## 0.0.0-bollards-20230520142731
+
+### Patch Changes
+
+- new message
+- Updated dependencies
+  - @business-as-code/address@0.0.0-bollards-20230520142731
+  - @business-as-code/core@0.0.0-bollards-20230520142731
+  - @business-as-code/error@0.0.0-bollards-20230520142731
+  - @business-as-code/fslib@0.0.0-bollards-20230520142731
+
+## 0.0.0-bollards-20230520142605
+
+### Patch Changes
+
+- new message
+- Updated dependencies
+  - @business-as-code/address@0.0.0-bollards-20230520142605
+  - @business-as-code/core@0.0.0-bollards-20230520142605
+  - @business-as-code/error@0.0.0-bollards-20230520142605
+  - @business-as-code/fslib@0.0.0-bollards-20230520142605
+
+## 0.0.0-bollards-20230520142317
+
+### Patch Changes
+
+- new message
+- Updated dependencies
+  - @business-as-code/address@0.0.0-bollards-20230520142317
+  - @business-as-code/core@0.0.0-bollards-20230520142317
+  - @business-as-code/error@0.0.0-bollards-20230520142317
+  - @business-as-code/fslib@0.0.0-bollards-20230520142317
+
 ## 0.0.0-bollards-20230520081808
 
 ### Patch Changes